+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased (develop)
 
 - added: Search bar to `EarnScene`
+- added: Price chart to `TransactionListScene`
 - added: Add Unizen DEX
+- changed: `TransactionListScene` split into two scenes: `TransactionListScene` and `TransactionListScene2`
 
 ## 4.21.1 (2025-01-28)